structure folders change
