+++ package.json
@@ -11,6 +11,7 @@
     "ionicons": "^7.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
     "react-table": "^7.8.0",
     "web-vitals": "^2.1.4"
