+++ src/components/Login.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Login = () => {
+		return (
+				<>
+						<h1>LOGIN</h1>
+				</>
+		);
+};
+
+export default Login;
