+++ src/index.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import SpeechToText from './App';
 import reportWebVitals from './reportWebVitals';
+import App from './App';
+import './index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <SpeechToText />
+    <App />
   </React.StrictMode>
 );
